Rename the project to `pyproject-hooks`
This more accurately reflects what this project is for.
