+++ .bumpversion.cfg
@@ -3,4 +3,4 @@ current_version = 0.13.0
 commit = True
 tag = True
 
-[bumpversion:file:src/pep517/__init__.py]
+[bumpversion:file:src/pyproject_hooks/__init__.py]
