+++ README.rst
@@ -23,7 +23,7 @@ Usage—you are responsible for ensuring build requirements are available:
 
     import os
     import tomli
-    from pep517.wrappers import Pep517HookCaller
+    from pyproject_hooks.wrappers import Pep517HookCaller
 
     src = 'path/to/source'  # Folder containing 'pyproject.toml'
     with open(os.path.join(src, 'pyproject.toml'), 'rb') as f:
@@ -34,7 +34,7 @@ Usage—you are responsible for ensuring build requirements are available:
     # an environment where they are available.
 
     hooks = Pep517HookCaller(
-        src, 
+        src,
         build_backend=build_sys['build-backend'],
         backend_path=build_sys.get('backend-path'),
     )
