+++ pyproject.toml
@@ -3,10 +3,11 @@ requires = ["flit_core >=3.2,<4"]
 build-backend = "flit_core.buildapi"
 
 [tool.flit.metadata]
-module = "pep517"
+module = "pyproject_hooks"
+dist-name = "pyproject-hooks"
 author = "Thomas Kluyver"
 author-email = "thomas@kluyver.me.uk"
-home-page = "https://github.com/pypa/pep517"
+home-page = "https://github.com/pypa/pyproject-hooks"
 description-file = "README.rst"
 requires = [
     "tomli >=1.1.0;python_version<'3.11'",
