+++ tests/test_hook_fallbacks.py
@@ -6,7 +6,7 @@ import tomli
 from testpath import assert_isfile, modified_env
 from testpath.tempdir import TemporaryDirectory
 
-from pep517.wrappers import HookMissing, Pep517HookCaller
+from pyproject_hooks.wrappers import HookMissing, Pep517HookCaller
 
 SAMPLES_DIR = pjoin(dirname(abspath(__file__)), 'samples')
 BUILDSYS_PKGS = pjoin(SAMPLES_DIR, 'buildsys_pkgs')
